# Conflicts:
#	Jenkinsfile
#	pom.xml

+++ core/src/main/java/org/gbif/vocabulary/service/VocabularyService.java
@@ -50,7 +50,10 @@ public interface VocabularyService extends BaseService<Vocabulary> {
    * @param deprecateConcepts if true the concepts of the vocabulary will be deprecated too
    */
   void deprecate(
-      int key, @NotBlank String deprecatedBy, @Nullable Integer replacementKey, boolean deprecateConcepts);
+      long key,
+      @NotBlank String deprecatedBy,
+      @Nullable Long replacementKey,
+      boolean deprecateConcepts);
 
   /**
    * Deprecates a concept without replacement.
@@ -59,7 +62,8 @@ public interface VocabularyService extends BaseService<Vocabulary> {
    * @param deprecatedBy name of the actor who deprecates the concept
    * @param deprecateConcepts if true the concepts of the concept will be deprecated too
    */
-  void deprecateWithoutReplacement(int key, @NotBlank String deprecatedBy, boolean deprecateConcepts);
+  void deprecateWithoutReplacement(
+      long key, @NotBlank String deprecatedBy, boolean deprecateConcepts);
 
   /**
    * Restores a deprecated concept.
@@ -67,5 +71,5 @@ public interface VocabularyService extends BaseService<Vocabulary> {
    * @param key key of the concept to undeprecate.
    * @param restoreDeprecatedConcepts if true it restores the deprecated concepts of the vocabulary
    */
-  void restoreDeprecated(int key, boolean restoreDeprecatedConcepts);
+  void restoreDeprecated(long key, boolean restoreDeprecatedConcepts);
 }
